Enable adding members after group creation
