+++ src/components/ui/loading.tsx
@@ -1,6 +1,6 @@
-const Loading = () => {
+const Loading = ({ className }: { className?: string }) => {
   return (
-    <div className="flex justify-center">
+    <div className={`flex justify-center ${className}`}>
       <div className="w-12 h-12 rounded-full animate-spin border-8 border-solid border-blue border-t-transparent" />
     </div>
   );
